Add files via upload
This project mainly focuses on analyzing the data which is provided by the respective company. My task is to derive insights and answer the questions asked by different departments. So that these insights are used by the operation team, support team, marketing team, etc to predict the overall growth or decline of a company’s fortune. So that the company gets better automation, better understanding between cross-functional teams, and more effective workflows.
In case study 1, there is a job_data table; in case study 2 there are users, events, and email_events tables.
In Case Study 1 the insights are found based on the following questions:
(i)	Number of jobs reviewed: Amount of jobs reviewed over time.
(ii)	Throughput: Calculate the 7-day rolling average of throughput. For throughput, do you prefer daily metric or 7-day rolling and why?
(iii)	Percentage share of each language: Share of each language for different contents.
(iv)	Duplicate rows: Rows that have the same value present in them

In case study 2 the insights are found based on the following questions:
(i)	User engagement: To measure the activeness of a user. Measuring if the user finds quality in a product/service.
(ii)	User Growth: A product’s users grow over time.
(iii)	Weekly Retention: Users are retained weekly after signing up for a product.
(iv)	Weekly Engagement: To measure the activeness of a user. Measuring if the user finds quality in a product/service weekly.
(v)	Email Engagement: Users engaging with the email service.

Using operation analytics, a business can determine which areas must be improved and uncover areas in which it can save money. Since this is one of a company’s most critical components, it can also predict the future profitability of the business.